pyproject.toml
[tool.poetry]
name = "cbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
ccxt = "^1.64.64"
DateTime = "^4.3"
matrix-nio = "^0.18.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
